Rename MailerProcessor to MailerSubscriber
Since it's including Spree::Event::Subscriber makes sense to use the
"Subscriber" suffix for the constant name.
